OTTER-613: share collapsible submitted code section
