+++ src/app/[orgSlug]/study/[studyId]/review/post-feedback-view.tsx
@@ -4,23 +4,17 @@ import { PageBreadcrumbs } from '@/components/page-breadcrumbs'
 import type { ReviewDecision } from '@/database/types'
 import { FeedbackAndNotesSection } from '@/components/study/feedback-and-notes'
 import { ProposalRequest } from '@/components/study/proposal-initial-request'
-import { ProposalStepHeader } from '@/components/study/proposal-step-header'
 import { Routes } from '@/lib/routes'
 import { STATUS_BANNER_BG } from '@/lib/status-banner-colors'
 import { type Submitted } from '@/schema/study'
-import { Box, Button, Collapse, Group, Stack, Text, Title } from '@mantine/core'
+import { Box, Button, Group, Stack, Text, Title } from '@mantine/core'
 import { CaretLeftIcon } from '@phosphor-icons/react'
 import { useRouter } from 'next/navigation'
 import type { Route } from 'next'
-import { useCallback, useRef, type ReactNode, type RefObject } from 'react'
+import type { ReactNode } from 'react'
 import type { CodeReviewFeedbackEntry, ProposalFeedbackEntry, SelectedStudy } from '@/server/actions/study.actions'
 import type { JobScanResult, LatestJobForStudy, StudyReviewWithMeta } from '@/server/db/queries'
-import { SubmittedCodeSection } from './submitted-code-section'
-import {
-    FULL_STUDY_CODE_TOGGLE_LABELS,
-    StudyCodeToggle,
-    useExpandable,
-} from '@/app/[orgSlug]/study/[studyId]/view/study-code-collapse'
+import { CollapsibleSubmittedCodeSection } from './collapsible-submitted-code-section'
 
 export type PostFeedbackKind = 'PROPOSAL' | 'CODE'
 
@@ -168,124 +162,6 @@ function PreviousButton({ href }: { href: Route }) {
     )
 }
 
-type SubmittedCodePanelProps = {
-    orgSlug: string
-    study: Submitted<SelectedStudy>
-    job: LatestJobForStudy | null
-    review: StudyReviewWithMeta | null
-    scan: JobScanResult | null
-    expanded: boolean
-    onCollapse: () => void
-    panelRef: RefObject<HTMLDivElement | null>
-}
-
-// The full "Submitted code" section (datasets, AI summary, security scan log, code viewer) is the
-// same one shown during active review. Per OTTER-613 the entire card is collapsed behind the
-// "View full study code" toggle in the step card, preserving its state between expansions.
-function SubmittedCodePanel({
-    orgSlug,
-    study,
-    job,
-    review,
-    scan,
-    expanded,
-    onCollapse,
-    panelRef,
-}: SubmittedCodePanelProps) {
-    // scan-presence is coupled to job-presence: the caller fetches both together and
-    // jobScanResultForJob never returns null (it falls back to {status:'IN-PROGRESS', logFile:null}),
-    // so scan is null exactly when job is null. The !scan check is the type-narrowing that lets us
-    // pass a non-null scan to SubmittedCodeSection; in practice it only fires on the null-job branch.
-    if (!job || !scan) return null
-    return (
-        <Collapse in={expanded} keepMounted>
-            <Box ref={panelRef} tabIndex={-1}>
-                <SubmittedCodeSection
-                    orgSlug={orgSlug}
-                    study={study}
-                    job={job}
-                    review={review}
-                    scan={scan}
-                    onCollapse={onCollapse}
-                />
-            </Box>
-        </Collapse>
-    )
-}
-
-type CodeSectionProps = {
-    isVisible: boolean
-    orgSlug: string
-    study: Submitted<SelectedStudy>
-    job: LatestJobForStudy | null
-    review: StudyReviewWithMeta | null
-    scan: JobScanResult | null
-    kindCopy: KindCopy
-    timestampLabel: string
-    timestampDate: Date | string | null
-    banner: ReactNode
-}
-
-function CodeSection({
-    isVisible,
-    orgSlug,
-    study,
-    job,
-    review,
-    scan,
-    kindCopy,
-    timestampLabel,
-    timestampDate,
-    banner,
-}: CodeSectionProps) {
-    const { expanded, toggle, collapse } = useExpandable()
-    const openerRef = useRef<HTMLButtonElement>(null)
-    const panelRef = useRef<HTMLDivElement>(null)
-    const onCollapse = useCallback(() => {
-        collapse()
-        requestAnimationFrame(() => openerRef.current?.focus())
-    }, [collapse])
-    const onExpand = useCallback(() => {
-        toggle()
-        requestAnimationFrame(() => panelRef.current?.focus())
-    }, [toggle])
-    if (!isVisible) return null
-    // Only offer the opener when there is a Submitted code panel behind it. SubmittedCodePanel
-    // returns null without a job/scan (e.g. the fallback auto-approved page), so an unconditional
-    // opener would otherwise expand to an empty card with no way back.
-    const hasSubmittedCode = Boolean(job && scan)
-    return (
-        <>
-            <ProposalStepHeader
-                stepLabel={kindCopy.stepLabel}
-                heading={kindCopy.heading}
-                studyTitle={study.title}
-                timestampDate={timestampDate}
-                timestampLabel={timestampLabel}
-                banner={banner}
-            >
-                <StudyCodeToggle
-                    ref={openerRef}
-                    isVisible={!expanded && hasSubmittedCode}
-                    expanded={false}
-                    onClick={onExpand}
-                    labels={FULL_STUDY_CODE_TOGGLE_LABELS}
-                />
-            </ProposalStepHeader>
-            <SubmittedCodePanel
-                orgSlug={orgSlug}
-                study={study}
-                job={job}
-                review={review}
-                scan={scan}
-                expanded={expanded}
-                onCollapse={onCollapse}
-                panelRef={panelRef}
-            />
-        </>
-    )
-}
-
 type ProposalSectionProps = {
     isVisible: boolean
     study: Submitted<SelectedStudy>
@@ -371,14 +247,15 @@ export function PostFeedbackView({
                 <Title order={1} fz={40} fw={700}>
                     Study proposal
                 </Title>
-                <CodeSection
+                <CollapsibleSubmittedCodeSection
                     isVisible={isCode}
                     orgSlug={orgSlug}
                     study={study}
                     job={job}
                     review={review}
                     scan={scan}
-                    kindCopy={kindCopy}
+                    stepLabel={kindCopy.stepLabel}
+                    heading={kindCopy.heading}
                     timestampLabel={timestampLabel}
                     timestampDate={timestampDate}
                     banner={banner}
